move timer functions to seperate file
I left the handler function because that's like the main function for
operation
