+++ main.c
@@ -4,15 +4,16 @@ SmotBot garden control unit
 */
 
 #include <stdio.h>
-#include <time.h> 
 #include <unistd.h>
-#include <signal.h>
 #include <string.h>
-#include <sys/time.h>
 #include <stdlib.h>
 #include <curl/curl.h>
 #include "main.h"
 #include "./inih/ini.h"
+#include "smot_timers.h"
+
+
+
 
 
 int main(int argc, char *argv[]) {
@@ -90,41 +91,6 @@ static int ini_handler_func(void *user, const char *section, const char *name, c
 	return 0;
 }
 
-//TODO change to accept prevous itimerval struct as argument
-int open_timer(int delay_number) {
-	struct itimerval timer;
-	
-	//TODO change 10 back to 60. 10 is for super fast
-	/* Configure the timer to expire after some minutes... */
-	timer.it_value.tv_sec = 10*delay_number;
-	timer.it_value.tv_usec = 0;
-	timer.it_interval.tv_sec = 10*delay_number;
-	timer.it_interval.tv_usec = 0;
-	
-	/* Start a virtual timer. It counts down whenever this process is executing. */
-	if(setitimer(ITIMER_VIRTUAL, &timer, NULL) != 0) {
-		printf("ERROR: opening timer failed\n");
-		return -1; //TODO I'd rather have only one final return
-	}
-	
-	return 0;
-}
-
-//TODO change to accept prevous sigaction struct as argument
-int attach_handler(void) {
-	//assigns an action to occur when timer expires
-	struct sigaction sa;
-	memset(&sa, 0, sizeof(sa)); //I don't actaully know what this does, but it's supposed to be used
-	
-	//calls the referanced function
-	sa.sa_handler = &timer_handler;
-	if(sigaction(SIGVTALRM, &sa, NULL) != 0) { 
-		printf("ERROR: handler failed\n");
-		return -1; //TODO I'd rather have only one final return
-	}
-	
-	return 0;
-}
 
 //####################&&&&&&&&&&&&&&&######################&&&&&&&&&&&&&############# bookmark
 void timer_handler(int signum) {
@@ -133,7 +99,7 @@ void timer_handler(int signum) {
 	
 	//if the current interval count is a prechosen time, preform approriate action
 	if((count == 0) || (count == (3*FULL_DAY))) {
-		int irl_count = get_irl_time();
+		int irl_count = get_irl_time(settings.increment_size);
 		if(irl_count < 0) {
 			printf("ERROR: system clock time failed\n");
 			result = -1;
@@ -191,23 +157,6 @@ void timer_handler(int signum) {
 	*/
 }
 
-int get_irl_time(void) {
-	time_t curtime;
-	struct tm *loc_time;
-	int intervals;
-
-	//Getting current time of system
-	curtime = time(NULL);
-
-	// Converting current time to local time, not seconds since Jan 1, 1970
-	loc_time = localtime(&curtime);
-	
-	//Convert to increments
-	intervals = FULL_HOUR*(loc_time->tm_hour) + (int)((loc_time->tm_min)/FULL_INC);
-	
-	//TODO add error handling
-	return intervals;
-}
 
 //reading sensor functions
 int read_dht22(void) {
