+++ smot_timers.h
@@ -1,0 +1,7 @@
+
+int open_timer(int);
+int attach_handler(void);
+void timer_handler(int);
+int get_irl_time(unsigned int);
+
+
